moves tracing packge into /infra
